requirements.txt
python-telegram-bot
keras
pillow
opencv-python
tensorflow
h5py
numpy
